Updated aux kicker code to work with two-part break beam. Can be programmed and can receive commands.
There was an issue that programmed boards could not receive commands,
which was traced back to changes in the bemixnet.c code regarding
checksums, which was not working. Perhaps the checksum was off on each
packet, which caused the aux kicker boards to discard everything. Also
added to the no_false_positives.c code was a grace period (of 5 main
loop iterations) for the break beam receiver to react after the LED
turned on. This was to resolve the issue that the robot would kick on a
break beam command even when the break beam was seemingly not broken.