Added hysteresis in breakbeam kick and increased lag after kick
We noticed that the robot would kick when it seemed like the LED did
not toggle indicating that the break beam was broken. We hypothesized
that this was due to the break beam being broken for a split second
causing the kick. We thus added a hysteresis variable in the break beam
code in which the break beam must be broken for HYSTERESIS_VALUE number
of main loops before the robot considers it triggered.

We also noticed funny noises after the robot kicked once, causing the
robot to not be able to charge up the capacitors again. This is most
likely due to the SCRs on the kicker board not turning off immediately
after kicking before the next instruction to charge the capacitors,
causing the kicker board to charge directly into the solenoid. We thus
increased the lag time after a kick from 125*256 instruction cycles to
256*256 instruction cycles.