Breakbeam Fixes in Aux Kicker Code
Changed timing of LED emitter in break beam to match datasheet spec.
Added counter (breakbeam_count_outer) to reset the break beam every ~4
sec in case it becomes stuck. Changed break beam trigger condition to
consider the break beam as not broken if it is ever not broken in a
burst period. Reduced hysteresis value to account for decreased burst
period (may not need hysteresis anymore due to combination of trigger
condition and grace period).