Added Brushless Firmware to Test Encoder
This firmware lets us test how many counts the encoder is putting out
per revolution of the wheel, in order to determine whether the encoder
is dirty/broken without having to first remove the encoder assembly.

To use, open MPLAB and in the Brushless project, replace the main_v2.c
file with main_v2_encodertest.c (remove the main_v2.c file from the
project and add main_v2_encodertest.c). Then click on the hammer/broom
icon at the top and select "Clean and Build for Debugging." Then
connect the programmer to the brushless board inserted in the position
corresponding to the motor you want to test, and turn on the robot.
Click on the download to chip icon at the top and select "Program
Device for Debugging."

When the programmer finishes downloading to the brushless board, click
the Debug menu on the toolbar and select "Debug Project." Do not move
the wheel as the program is running, until the program stops at the
breakpoint. Mouse-over the POSCNT variable -- you should see the value
0b1000000000000000. (Or something close to that. If not, repeat the
steps in this paragraph.) Get ready. Click the Run button and
immedately rotate the wheel clockwise for 1 rotation. Wait until the
program reaches the breakpoint again, and mouse-over the POSCNT
variable. You should see something greater than the original value. (If
not, repeat the steps in this paragraph, instead turning the wheel in
the opposite direction.) If you were too slow, repeat the steps in this
paragraph.

Take the difference between the final value and original value and
convert this to decimal from binary (you can use Google to do that).
The number should be roughly 3000. Repeat a couple times for
consistency. Test the other wheels using the same procedure and compare
the counts.

When finished, you need to reprogram the brushless board with the
original code in order to run normally. To do so, first click the Stop
button if the debug session is currently running. Replace the
main_v2_encodertest.c file in the project with the main_v2.c file.
Click the hammer/broom icon and select "Clean and Build Project" (NOT
for debugging). Connect the programmer to the brushless board and turn
the robot on. Click the download to chip icon and select "Make and
Program Device" (NOT for debugging).

Note: one of the wheels corresponds to a brushless board position that
is very hard to connect a programmer to (because the programmer
intersects one of the capacitors). One of the brushless boards has its
programming pins bent up, and this brushless board works well in that
position. Otherwise, CAREFULLY bend up the programming pins on another
brushless board using a set of pliers.